chore(testrefactor): make use of correct test folder structure and refactor test -[Delivers #161682143]
