+++ src/components/authors/AuthorForm/__tests__/AuthorForm.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import AuthorForm from './AuthorForm';
+import AuthorForm from '../AuthorForm';
 
 describe('AboutPage Component', () => {
 	const props = {
@@ -18,13 +18,19 @@ describe('AboutPage Component', () => {
 			lastName: ''
 		}
 	};
-  
+
 	const wrapper = shallow(<AuthorForm {...props} />);
 
 	it('should render correctly', () => {
+		expect(wrapper.exists()).toBeTruthy();
 		expect(wrapper).toMatchSnapshot();
+	});
+
+	it('should render form, TextInput, button', () => {
 		expect(wrapper.find('form')).toHaveLength(1);
 		expect(wrapper.find('button')).toHaveLength(1);
 		expect(wrapper.find('button').text()).toEqual('Save');
+		expect(wrapper.find('TextInput')).toHaveLength(2);
+		expect(wrapper.find('TextInput')).toHaveLength(2);
 	});
 });
