+++ src/components/common/TextInput/__tests__/TextInput.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import TextInput from './TextInput';
+import TextInput from '../TextInput';
 
 describe('TextInput Component', () => {
 	const props = {
