+++ src/components/notFoundPage/__tests__/NotFound.test.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import NotFound from './NotFound';
+import NotFound from '../NotFound';
 
 
 describe('Header Component', () => {
 	const wrapper = shallow(<NotFound />);
-  
+
 	it('should render correctly', () => {
 		expect(wrapper).toMatchSnapshot();
 	});
