+++ src/reducers/__tests__/courseReducer.test.js
@@ -1,21 +1,21 @@
-import coursesReducer from './courseReducer';
+import coursesReducer from '../courseReducer';
 import {
 	CREATE_COURSE_SUCCESS,
 	LOAD_COURSES_SUCCESS,
 	UPDATE_COURSE_SUCCESS,
 	DELETE_COURSE_SUCCESS
-} from '../actions/actionTypes';
-import { initialCourses } from './initialState';
-import courses from '../__test__/__mocks__/mockCourses';
-import { 
+} from '../../actions/actionTypes';
+import { initialCourses } from '../initialState';
+import courses from '../../__test__/__mocks__/mockCourses';
+import {
 	createCourseSuccess, loadCoursesSuccess
-} from '../actions/courseActions';
+} from '../../actions/courseActions';
 
 describe('Course Reducer', () => {
 	it('should return initial state', () => {
 		expect(coursesReducer(undefined, { courses: [] })).toEqual(initialCourses);
 	});
-  
+
 	it('should add course when passed CREATE_COURSE_SUCCESS', () => {
 		const action = {
 			type: CREATE_COURSE_SUCCESS,
@@ -40,7 +40,7 @@ describe('Course Reducer', () => {
 			type: UPDATE_COURSE_SUCCESS,
 			course: courses[0]
 		};
-		
+
 		initialCourses.courses.push(courses[0]);
 		expect(coursesReducer(initialCourses, action))
 			.toEqual({ courses: [courses[0]], isLoading: false });
@@ -51,7 +51,7 @@ describe('Course Reducer', () => {
 			type: DELETE_COURSE_SUCCESS,
 			courseId: courses[0].id
 		};
-		
+
 		initialCourses.courses.push(courses[0]);
 		expect(coursesReducer(initialCourses, action))
 			.toEqual({ courses: [], isLoading: false });
